int -> size_t

+++ libft/ft_itoa.c
@@ -6,15 +6,15 @@
 /*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 16:12:02 by marvin            #+#    #+#             */
-/*   Updated: 2022/11/11 21:57:22 by keitakah         ###   ########.fr       */
+/*   Updated: 2022/11/13 11:56:18 by keitakah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	str_cnt(int n)
+static size_t	str_cnt(int n)
 {
-	int	cnt;
+	size_t	cnt;
 
 	cnt = 0;
 	if (n == 0)
@@ -31,8 +31,8 @@ static int	str_cnt(int n)
 
 static char	*str_min_fit(char *str)
 {
-	int		i;
-	char	*min;
+	size_t		i;
+	char		*min;
 
 	i = 0;
 	min = "-2147483648";
@@ -43,7 +43,7 @@ static char	*str_min_fit(char *str)
 
 static char	*str_fit(char *str, int n)
 {
-	int	i;
+	size_t	i;
 
 	i = str_cnt(n);
 	if (n == 0)
